backend/requirements.txt
aniso8601==7.0.0
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
entrypoints==0.3
fastapi==0.54.1
fastdiff==0.2.0
flake8==3.7.9
graphene==2.1.8
graphql-core==2.3.1
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
idna==2.9
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pathspec==0.8.0
pluggy==0.13.1
promise==2.3
py==1.8.1
pycodestyle==2.5.0
pydantic==1.5.1
pyflakes==2.1.1
pyparsing==2.4.7
pytest==5.4.2
regex==2020.5.7
requests==2.23.0
Rx==1.6.1
six==1.14.0
snapshottest==0.5.1
SQLAlchemy==1.3.16
starlette==0.13.2
termcolor==1.1.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
wasmer==0.4.1
wcwidth==0.1.9
websockets==8.1
